--- VissimSimulator/VissimSimulator/Program.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace VISSIMSimulator
-{
-    public class Program
-    {
-        public static void Main()
-        {
-            EventSimulator simulator = new EventSimulator();
-            simulator.Run();
-        }
-    }
-}
